+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 - fixed deferred table copy when task is invalid (introduced with 0.2.2)
 - fixed mssql to not reflect full schema while renamings happen
 - fixed clearing of metadata tables for lazy tables and raw sql tables
+- fixed mssql synonym resolution when reading input table for pandas task
 
 0.2.2 (2023-03-31)
 ------------------